Add experimental-TAK server and datachannel_tak module
This commit adds a new experimental server, "TAK", and introduces a new module called "datachannel_tak". It includes all the necessary files, from package JSON and configuration files, to the core JavaScript files for running the server and module. The TAK server uses the Docker system for creating and managing containers, which it pulls from a specific image. The Datachannel_tak module includes a manufacture data that interacts with GraphQL.