Refactor handling of user roles and improve token signing (#167)
Updated treatment of user roles in the context of organizations in various parts of the codebase. This included clarifications in the README, new TypeScript types in the 'user_credentials_cache' app, among other changes. Token signing in the 'tokens' action in the 'catalyst-ui' app was also enhanced to include the 'cfToken'.