Some Minor fixes (#165)
### TL;DR

This pull request removes some redundant code from the ChannelDetails and CreateChannelForm components and also cleans up package.json by removing a duplicate dependency.

### What changed?

In both ChannelDetails and CreateChannelForm components, the form control for 'organization' has been removed as it is not required. The Input field for 'accessSwitch' in ChannelDetails now uses default value instead of value property. Also, the '@cloudflare/workers-types' dependency was duplicated in the package.json of catalyst-ui, which has been rectified.