add Back button to partner detailed view (#294)
### TL;DR

Refactored the layout in PartnerDetailedComponent from ListView to DetailedView.

### What changed?

Replaced ListView with DetailedView in PartnerDetailedComponent. Removed the table wrapper around the OrbisTable component and directly inserted it into the DetailedView layout.
Before

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/gTJhzORipy1tCg65Vgvf/c8db28ee-74ab-4bfd-b72e-4d89753b6c08.png)


Now
![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/gTJhzORipy1tCg65Vgvf/e129de74-7fd5-43e8-b973-9e2292c9a8c9.png)


### How to test?

To test, navigate to the partner detail view in the Catalyst UI and verify that the layout appears correctly. The OrbisTable displaying the shared channels should now be directly in the DetailedView layout.

### Why make this change?

This change is made to ensure more clarity and linearity for our components. Using the DetailedView layout aligns better with our design principles and improves user experience.

---