Refactor the datachannel-next package and improve logs handling
This commit refactors the datachannel-next package and improves the handling of log messages. The major change is the conversion of the run.js file into a TypeScript file (run.ts). This new TypeScript file includes refactoring for better readability and introduces new error handling for TAK server disconnects. Additionally, the handling of log messages in the MessageLogCard component is improved by using ReactMarkdown to display messages and added a handling function for 'cot' socket events. The commit also modifies relevant package scripts to use the 'bun' module instead of 'node' for running the app. Finally, unneeded dependencies are removed from 'datachannel-tak' package.